+++ src/browser/modules/Stream/SysInfoFrame/LegacySysInfoFrame/sysinfoUtils.tsx
