+++ src/browser/modules/Stream/SysInfoFrame/sysinfoUtils.tsx
